Use list comprehension and lazy iteration
For MemInfo.search, in Python 3 filter returns a generator
object which is not a list (but can be iterated). We force
it to become a list by using list comprehensions which are
more expressive at the same time. The same can be done by
doing list(filter(...)) but that requires the list to be
iterated twice.

For MemInfo.get, we don't need to consume the whole file
before processing. We can do stream processing to avoid
needing to read the whole file into memory.